+++ lib/minfra/cli/env.rb
@@ -2,12 +2,13 @@ module Minfra
   module Cli
     class Env
       attr_reader :name, :hiera
-      def initialize(name:, hiera_root:,   hiera_env_path:)
+      def initialize(name:, hiera_root:,   hiera_env_path:, hiera_debug_lookups: )
         @name = name            
         @hiera = HieraLooker.new( 
           root: hiera_root,
           env_name: name,
-          env_path: hiera_env_path
+          env_path: hiera_env_path,
+          debug_lookups: hiera_debug_lookups
         )
       end