+++ lib/minfra/cli/hiera_looker.rb
@@ -1,9 +1,11 @@
 # frozen_string_literal: true
 
+require 'English'
 module Minfra
   module Cli
     class HieraLooker
-      def initialize(root:, env_name:, env_path:)
+      include Logging
+      def initialize(root:, env_name:, env_path:, debug_lookups: false)
         @root = Pathname.new(root)
         @env_name = env_name
         @cache = {}
@@ -18,13 +20,15 @@ module Minfra
         scope = { 'hieraroot' => @root.to_s, 'env' => @env_name }
 
         @special_lookups = @hiera.lookup('lookup_options', {}, scope, nil, :priority)
-        
+
         node_scope = @hiera.lookup('env', {}, scope, nil, :deeper)
         @scope = scope.merge(node_scope)
+        @debug_lookups = debug_lookups
       end
 
       def l(value, default = nil)
-#        debugger if @env_name == 'production-management' && value == 'env.tags'
+        debug "hiera: #{value}" if @debug_lookups
+        #        debugger if @env_name == 'production-management' && value == 'env.tags'
         return @cache[value] if @cache.key?(value)
 
         values = value.split('.')
@@ -35,7 +39,18 @@ module Minfra
                       else
                         :deep
                       end
-        result = @hiera.lookup(fst_value, default, @scope, nil, lookup_type)
+        begin
+          result = @hiera.lookup(fst_value, default, @scope, nil, lookup_type)
+        rescue GPGME::Error::NoSecretKey
+          error("Have no gpg configuration to decrypt your hiera key: #{value}")
+          raise Errors::ExitError
+        rescue GPGME::Error::BadPassphrase
+          error("Your password was wrong for hiera key: #{value}")
+          raise Errors::ExitError
+        rescue GPGME::Error
+          error("Having decrypt problems for hiera key: #{value}, #{$ERROR_INFO.message}")
+          raise Errors::ExitError
+        end
         result = result.dig(*values) if !values.empty? && result.is_a?(Hash) # we return nil or the scalar value and only drill down on hashes
         result = default if result.nil?
         result = Hashie::Mash.new(result) if result.is_a?(Hash)
@@ -46,9 +61,9 @@ module Minfra
       def l!(value, default = nil)
         v = l(value, default)
         raise("Value not found! #{value}") if v.nil?
+
         v
       end
-      
     end
   end
 end