+++ lib/minfra/cli/logging.rb
@@ -4,7 +4,7 @@ module Minfra
   module Cli
     module Logging
       def error(str)
-        logger.error str
+        logger.error(colored(str, :red))
       end
 
       def exit_error(str)
@@ -17,11 +17,11 @@ module Minfra
       end
 
       def warn(str)
-        logger.warn str
+        logger.warn(colored(str, :yellow))
       end
 
       def debug(str)
-        logger.debug str
+        logger.debug(str)
       end
 
       def deprecated(comment)
@@ -32,6 +32,14 @@ module Minfra
 
       def logger
         Minfra::Cli.logger
+      end
+      LOGGING_COLORS={ red: '1;31;40', green: '1;32;40' , dark_green: '0;32;40' , yellow: '1;33;40' }
+      def colored(str, color)
+        if $stdout.isatty
+          "\e[#{LOGGING_COLORS[color] || LOGGING_COLORS[:color1]}m#{str}\e[0m"
+        else
+          str
+        end    
       end
     end
   end